requirements.txt
# Flask
# diffusers
# transformers
# torch
# PyPDF2

Flask==2.3.3
google-cloud-aiplatform==1.42.1
PyMuPDF==1.23.8
python-dotenv==1.0.0
Pillow==10.0.1
Werkzeug==2.3.7